Fix display of CONVOS_SOMETHING_SOMETHING names
In Markdown, a word between underscores will be in displayed in
_italics_.

Apparently, at least in the renderer that displays the documentation,
this does _not_ require a word bou_nda_ry, so also if these underscores
are inside a word, it will show the part inside the word in italics.
And for the configuration variables of Convos, such as CONVOS_LOG_FILE,
this causes a display issue because the underscores are _removed_ from
the word and the middle part is displayed in italics.

This commit changes all those places in the documentation to have those
variables in backticks. This has the added benefit that these are now
rendered in a nice monospaced font.